# server/api/openapi.yaml
info:
  version: 0.0.2
  title: "shifttree API"
openapi: 3.0.3
paths:
...
  /schedules/{scheduleId}:
    get:
      security:
        - Bearer: []
      summary: Get schedule
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Get detailed info about a schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleInfo"
        404:
          description: Schedule with specified ID does not exist, or user does not have access to it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      security:
        - Bearer: []
      summary: Delete schedule
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Schedule deleted
        404:
          description: Schedule with specified ID does not exist, or user does not have access to it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: User does not have permission to delete the schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      required:
        - error
      properties:
        error:
          type: string
        detail:
          type: string
...
    UserInfoPreview:
      type: object
      required:
        - id
        - displayName
        - email
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        email:
          type: string
        profileImageUrl:
          type: string
...
    ScheduleInfo:
      allOf:
        - $ref: "#/components/schemas/ScheduleInfoPreview"
    ScheduleInfoPreview:
      type: object
      required:
        - id
        - name
        - description
        - owner
        - startTime
        - endTime
        - role
        - state
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        owner:
          $ref: "#/components/schemas/UserInfoPreview"
        startTime:
          type: string
          format: date-time
          nullable: true
        endTime:
          type: string
          format: date-time
          nullable: true
        role:
          type: string
          enum:
            - owner
            - member
            - manager
        state:
          type: string
          enum:
            - open
            - closed
...